Minor update to ensure j1 remains integer.
git-svn-id: https://scm.physics.ox.ac.uk/svn/radtrancode/trunk@551 4dc0e162-f54d-4b9e-8a18-81c7a7f2ddce

+++ radtran/cirsrad/cirsrad_wave.f
@@ -996,7 +996,7 @@ C               matrix inversion crashing
 		enddo
 
 C               Check to see if this is a limb path
-                j1=0.5*nlays
+                j1=int(0.5*nlays)
 
                 p1=press(layinc(j1,ipath))
                 p2=press(layinc(nlays,ipath))
