+++ radtran/cirsrad/lblrad_wave.f
@@ -894,7 +894,7 @@ C     1                  ' output'
 		enddo
 
 C               Check to see if this is a limb path
-                j1=0.5*nlays
+                j1=int(0.5*nlays)
 
                 p1=press(layinc(j1,ipath))
                 p2=press(layinc(nlays,ipath))
